Crate mlx5

A mlx5 driver for a ConnectX-5 100GbE Network Interface Card.

Currently we only support reading the device PCI space, mapping the initialization segment, and setting up a command queue to pass commands to the NIC.

All information is taken from the Mellanox Adapters Programmer’s Reference Manual (PRM) Rev 0.54, unless otherwise specified.

Structs

ConnectX5Nic
Struct representing a ConnectX-5 network interface card.

Constants

CONNECTX5_DEV
Device ID for the ConnectX-5 NIC

MLX_VEND
Vendor ID for Mellanox

Functions

get_mlx5_nic
Returns a reference to the NIC wrapped in a MutexIrqSafe, if it exists and has been initialized.