Crate mlx5

A mlx5 driver for a ConnectX-5 100GbE Network Interface Card.

Currently the following steps are completed

reading the device PCI space and mapping the initialization segment

setting up a command queue to pass commands to the NIC

setting up a single send and receive queue

functions to send packets

All information is taken from the Mellanox Adapters Programmer’s Reference Manual (PRM) Rev 0.54, unless otherwise specified.

Structs

ConnectX5Nic

Struct representing a ConnectX-5 network interface card.

Constants

CONNECTX5_DEV

Device ID for the ConnectX-5 NIC

CONNECTX5_EX_DEV

Device ID for the ConnectX-5-EX NIC

MLX_VEND

Vendor ID for Mellanox

Functions

get_mlx5_nic

Returns a reference to the NIC wrapped in a MutexIrqSafe, if it exists and has been initialized.